refactor: Remove printMyBarcode module and related files
