--- src/config/PrinterList.json
@@ -1,10 +1,0 @@
-[
-  { "printerName": "g216bc" },
-  { "printerName": "h105bc" },
-  { "printerName": "stub" },
-  { "printerName": "aPrinterName" },
-  { "printerName": "ssrtubebc-sq1" },
-  { "printerName": "aa309bc1" },
-  { "printerName": "g311bc1" },
-  { "printerName": "aa309bc3" }
-]
