+++ src/store/index.js
@@ -61,26 +61,21 @@
 import { createStore } from 'vuex'
 import config from '@/api/Config'
 import build from '@/api/v1/ApiBuilder'
-import PrinterList from '@/config/PrinterList'
 import traction from '@/store/traction'
-import printMyBarcode from '@/store/printMyBarcode'
 import PlateMap from '@/config/PlateMap'
 
 const store = createStore({
   state: {
     api: build({ config }),
-    printers: PrinterList,
     plateMap: PlateMap,
   },
   mutations: {},
   getters: {
     api: (state) => state.api,
-    printers: (state) => state.printers.map((obj) => obj.printerName),
     plateMap: (state) => state.plateMap,
   },
   modules: {
     traction,
-    printMyBarcode,
   },
 })
