--- src/store/printMyBarcode/getters.js
@@ -1,5 +1,0 @@
-const getters = {
-  printJobRequest: (state, getters, rootState) => rootState.api.printMyBarcode.print_jobs,
-  tubeLabelTemplateName: (state) => state.tubeLabelTemplateName,
-}
-export default getters
