--- src/store/printMyBarcode/index.js
@@ -1,12 +1,0 @@
-import state from './state'
-import getters from './getters'
-import actions from './actions'
-
-const printMyBarcode = {
-  namespaced: true,
-  state,
-  getters,
-  actions,
-}
-
-export default printMyBarcode
