--- src/store/printMyBarcode/state.js
@@ -1,5 +1,0 @@
-const state = {
-  tubeLabelTemplateName: import.meta.env.VITE_TUBE_LABEL_TEMPLATE_NAME,
-}
-
-export default state
