+++ tests/unit/store/index.spec.js
@@ -7,10 +7,6 @@ describe('index', () => {
       expect(Store.state.api).toBeDefined()
     })
 
-    it('has a list of printers', () => {
-      expect(Store.state.printers).toBeDefined()
-    })
-
     describe('api', () => {
       it('contains multiple resources', () => {
         expect(Store.state.api.traction).toBeDefined()
@@ -30,10 +26,6 @@ describe('index', () => {
   describe('getters', () => {
     it('api', () => {
       expect(Store.getters.api).toBeDefined()
-    })
-
-    it('printers', () => {
-      expect(Store.getters.printers).toBeDefined()
     })
 
     it('PlateMap', () => {
